refactor counterfactual list
